Rename the CRD plugin package to simply crd
