+++ plugins/crd/doc.go
@@ -1,2 +1,2 @@
 // Package netmeshplugincrd implements the CRD plugin for networkservicemesh
-package netmeshplugincrd
+package crd
