+++ plugins/crd/plugin_impl_crd.go
@@ -14,7 +14,7 @@
 
 // //go:generate protoc -I ./model/pod --go_out=plugins=grpc:./model/pod ./model/pod/pod.proto
 
-package netmeshplugincrd
+package crd
 
 import (
 	"fmt"
